Copy brief response questions and manifests to dos2 from dos1
As we had been creating the 'new supplier flow' we only made
the changes to the dos1 content. The dos2 content will need to
reflect the new flow also.

Note, I have not copied across the legacy brief response questions
or manifests. This however relies on dos2 not being live until at
least 2 weeks after the new supplier flow gets released. If this
were not the case then we would also need to have the legacy
manifests and questions in the dos2 content.

Further note, if there are any more changes to dos1 brief response
content then we will need to update the dos2 content also to
reflect.